přesun pro editaci a zobrazeni statistik do managera evidence
